+++ GIT_STRATEGY.md
@@ -9,7 +9,7 @@
 ### Layer 1: 영구 브랜치 (Permanent)
 
 - **`main`**: 상용(Production) 서버에 배포되는 최상위 신뢰 브랜치입니다.
-- **`develop`**: 모든 개발 작업이 통합되는 기준 브랜치입니다. 스테이징 환경 배포의 기준이 됩니다.
+- **`dev`**: 모든 개발 작업이 통합되는 기준 브랜치입니다. 스테이징 환경 배포의 기준이 됩니다.
 
 ### Layer 2: 카테고리 (Category)
